Add files via upload
hey guys, I could not figure out the graphics but while I was hunting around for examples but while I was hunting around I came across this gem. I've taken the liberty of designing the level based on the picture that Professor Lau had uploaded its about half of the map A,B and c building. anyways you can run it is VS 2019 and have a look. If we can learn something from it then I'm happy I could contribute.